Revert "Python 3.x compatability - basic generation and rendering now works, some rendering still needs fixing"
This reverts commit dd0b23b3a1d757bf004f692fbd924dc293c62df0.

+++ drawUtils.py
@@ -49,7 +49,7 @@ def drawSquare(v0, width, color, blend=False):
 def drawRivers(verts, arrowHeads=False):
     if len(verts) > 0:
         pyglet.gl.glColor4f(0.0, 0.1, 1.0, 1.0)
-        numVerts = int(len(verts)/2)
+        numVerts = len(verts)/2
         pyglet.graphics.draw(numVerts, pyglet.gl.GL_LINES,
             ('v2f', verts)
         )
